requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contextlib2==21.6.0
cryptography==43.0.1
Django==5.1.2
django-cors-headers==4.5.0
django-ninja==1.3.0
django-ninja-extra==0.21.4
django-ninja-jwt==5.3.4
django-prometheus==2.3.1
h11==0.14.0
idna==3.10
injector==0.22.0
prometheus_client==0.21.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
